Refactored Branch to new Folder Structure
