+++ src/App/DisplayElements/DisplayElements.js
@@ -1,4 +1,6 @@
 import TextDisplay from "./TextDisplay/TextDisplay";
+import ImageDisplay from "./ImageDisplay/ImageDisplay";
 export const DisplayElements = {
     'textdisplay':TextDisplay,
+    "imagedisplay": ImageDisplay,
 }
