+++ src/App/UXComponents/PropertyPanel/PropertiesMenu.css
